requirements.txt
Flask==1.0.2
Jinja2==2.10
beautifulsoup4==4.7.1
requests==2.21.0
tweepy
google-cloud-language
nltk
flask_sqlalchemy
